Camera can now see through arena if outside
Lots of things had to be implemented:
- System for declaring transparency for a game object geometry
- System to update geometry buffers based on world state
- System for color scaling in shaders

Because transparency objects need to be painted last, render currently uses a
hack that draws world.objects in reverse, on the assumption that the arena is
the first object in the list, and that it's the only one with transparency
elements.

Clean-up of code incoming.